Add module terraform-ovh-patroni_cluster
This module aim to abstract all the underlying resources deployment
necessary to build a patroni cluster.

Side note: terraform validate fail to correctly the module
```
❯ terraform plan | grep Plan
Plan: 133 to add, 0 to change, 0 to destroy.
❯ terraform validate
Success! The configuration is valid.

❯ terraform validate modules/terraform-ovh-patroni_cluster

Error: Missing required argument

The argument "endpoint" is required, but was not set.

❯ vim modules/terraform-ovh-patroni_cluster/providers.tf
❯ cat modules/terraform-ovh-patroni_cluster/providers.tf
provider "ovh" {
  endpoint = var.ovh_endpoint
}
❯ terraform validate modules/terraform-ovh-patroni_cluster
Success! The configuration is valid.

❯ terraform validate

Error: Module does not support count

  on patroni.tf line 16, in module "patroni-data-node-shard":
  16:   count = 3

Module "patroni-data-node-shard" cannot be used with count because it contains
a nested provider configuration for "ovh", at
modules/terraform-ovh-patroni_cluster/providers.tf:1,10-15.

This module can be made compatible with count by changing it to receive all of
its provider configurations from the calling module, by using the "providers"
argument in the calling module block.
```

See:
* https://github.com/hashicorp/terraform/issues/21408
* https://github.com/hashicorp/terraform/pull/24896

The error above is correctly fixed in v0.15.0-alpha20210210 of
terraform.