[terragrunt] Fix attempt
With terragrunt we stumble upon on a HTTP412 error when managing the
remote state. This was due to the fact that all child plan were writing
in the same file hence ending in a race condition. To fix that we are
using the name of the directory to name the state file.

While doing so we have taken the opportunity to refactor the variables
and leverage tfvars files. Each child plan has it's own set of
variables. Terragrunt also takes care of deploying all common variables
in a file named `common.tf`.

Finally let's fix the `.terraform.lock.hcl`.

Having such big commit should be illegal.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>