[terraform] Include secgroup in child plan
Instead of having a single child plan responsible for deploying all
secgroups, let's split them where they matter. The main benefit to do so
is to **not** deploy more than enough if we deploy partially the
infrastructure. Another reason is to have everything related to a piece
of infrastructure in the same plan.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>