[postgresql] Include timescaledb-tune in patroni bootstrap parameters
This is a first step to correctly and automatically tune postgresql for
timescaledb workload. At the moment this is aiming for i1-180 instances
which are my go to flavor. This can be improve by computing the values
based on the specs of the instances.

See: https://github.com/timescale/timescaledb-tune

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>