Experimental auto annotations
When refining configuration and infrastructure it's hard to remember
what happen at a given time. Let's try top use grafana annotation for
this matter. Each playbook will POST an annotation with ~meaningful
information at the end of the play. It would be a good idea to find a
way to create an annotation when a service is reloaded, restarted or
simply stopped.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>