[prometheus] Add clustered configuration
By adding 4 prometheus instance it would be easier to trick the remote
storage and better assess it. This commit introduce 2 clusters:
* cluster 1, prometheus-1 & prometheus-2
* cluster 2, prometheus-3 & prometheus-4

Apart from the cluster id they have identical configuration and write to
the same remote storage. The cluster id will be used in the
file_sd_config to create more fake target.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>