[node_exporter] Set web_max_requests=0
The more we put pressure on the remote_storage the more we stress the
node_exporter running on tsbs. Specifically on the `collector.systemd`
which is responsible for executing system calls and therefore on the
underlying system.

A pull-request has been open to make the parameter change possible.
We will switch back to upstream once it is merged.

See: https://github.com/cloudalchemy/ansible-node-exporter/pull/249

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>