[prometheus] Revert from using cortex loadbalancer
At this point of the discovery the prometheus clusters are pushing so
much samples that the loadbalancer in front of Cortex distributor is
getting a hard time to keep up and are causing prometheus to lag.
Let's remove it from the write path so we can accumulate data and work
on the read path.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>